requirements.txt
fastapi
uvicorn
requests
python-dotenv
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Database
motor==3.3.1  # Async MongoDB driver
redis==5.0.1  # Redis for caching

# WhatsApp Integration
httpx==0.25.2  # Async HTTP client
python-multipart==0.0.6  # Form data handling

# AI Integration
google-generativeai==0.3.1  # Gemini AI
tiktoken==0.5.1  # Token counting for AI

# Security & Monitoring
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4  # Password hashing
sentry-sdk==1.39.1  # Error tracking
prometheus-fastapi-instrumentator==6.1.0  # Metrics

# Development Tools
black==23.11.0  # Code formatting
isort==5.12.0  # Import sorting
mypy==1.7.1  # Type checking
pytest==7.4.3  # Testing
pytest-asyncio==0.21.1  # Async testing
pytest-cov==4.1.0  # Coverage reporting
phonenumbers==8.13.12
